Add spec to validate service_message assertion
A big value with `perform_nlp!` is that memoizes the `nlp_result` between controllers during `step_to`. The reason this works is because `service_message` is passed between controllers and it itself contains the `nlp_result`. Other upcoming features will also utilize this so adding coverage to ensure this doesn’t regress.